+++ x/erc20/keeper/grpc_query_test.go
@@ -3,6 +3,7 @@ package keeper_test
 import (
 	"fmt"
 
+	testconstants "github.com/zenanetwork/zena/testutil/constants"
 	utiltx "github.com/zenanetwork/zena/testutil/tx"
 	"github.com/zenanetwork/zena/x/erc20/types"
 	exampleapp "github.com/zenanetwork/zena/zenad"
@@ -31,7 +32,7 @@ func (suite *KeeperTestSuite) TestTokenPairs() {
 					Pagination: &query.PageResponse{
 						Total: 1,
 					},
-					TokenPairs: exampleapp.ExampleTokenPairs,
+					TokenPairs: testconstants.ExampleTokenPairs,
 				}
 			},
 			true,
@@ -42,7 +43,7 @@ func (suite *KeeperTestSuite) TestTokenPairs() {
 				req = &types.QueryTokenPairsRequest{
 					Pagination: &query.PageRequest{Limit: 10, CountTotal: true},
 				}
-				pairs := exampleapp.ExampleTokenPairs
+				pairs := testconstants.ExampleTokenPairs
 				pair := types.NewTokenPair(utiltx.GenerateAddress(), "coin", types.OWNER_MODULE)
 				suite.network.App.Erc20Keeper.SetTokenPair(ctx, pair)
 				pairs = append(pairs, pair)
@@ -58,7 +59,7 @@ func (suite *KeeperTestSuite) TestTokenPairs() {
 			"2 pairs registered wo/pagination",
 			func() {
 				req = &types.QueryTokenPairsRequest{}
-				pairs := exampleapp.ExampleTokenPairs
+				pairs := testconstants.ExampleTokenPairs
 
 				pair := types.NewTokenPair(utiltx.GenerateAddress(), "coin", types.OWNER_MODULE)
 				pair2 := types.NewTokenPair(utiltx.GenerateAddress(), "coin2", types.OWNER_MODULE)